+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # Changelog
 
-## Unreleased
+## 5.7.1
 
 * Fix: Sentry Timber integration does not submit msg.formatted breadcrumbs (#1957)
 * Fix: ANR WatchDog won't crash on SecurityException ([#1962](https://github.com/getsentry/sentry-java/pull/1962))
-
 
 ## 5.7.0